refactor: split zig.rs into smaller modules
Convert the 3000-line src/zig.rs into a src/zig/ directory split by
concern, as a pure move with no behavior changes:

- mod.rs: the Zig clap subcommand and its execution entry points
  (zig cc/ar/dlltool dispatch), plus re-exports keeping every public
  path (zig::Zig, zig::ZigWrapper, zig::prepare_zig_linker,
  zig::adjust_canonicalization) unchanged for library consumers
- target_info.rs: TargetInfo and its target-triple predicates
- linker_args.rs: linker argument filtering and CPU/feature name
  mapping, with the whole test_filter_* suite alongside
- locate.rs: finding and validating the zig toolchain (find_zig,
  zig_version, lib_dir, cache_dir)
- cargo_env.rs: cargo build environment setup (apply_command_env,
  bindgen options, OS deps, cmake toolchain)
- wrapper.rs: linker wrapper script generation (prepare_zig_linker,
  TargetFlags, script quoting helpers)

Helpers that now cross file boundaries became pub(crate); nothing new
is pub.
